Move image_utils to types
