+++ tests/unittests/types/test_mage_utils.py
@@ -20,7 +20,7 @@ from pathlib import Path
 from PIL import Image as PILImage
 
 from towhee.types import Image
-from towhee.utils.image_utils import from_pil, to_pil
+from towhee.types.image_utils import from_pil, to_pil
 
 logo_path = os.path.join(Path(__file__).parent.parent.parent.parent.resolve(), 'towhee_logo.png')
